Tidy up some more
